add one api call

+++ sql/migrations/1605037011_create_user_cart.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE users_cart (
     product_id uuid REFERENCES products(id),
     user_id uuid REFERENCES users(id),
-    amount numeric,
+    amount INT,
     CONSTRAINT product_user_pkey PRIMARY KEY (product_id, user_id)
 );
